requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
colorama==0.4.4
commonmark==0.9.1
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==3.2.6
django-allauth==0.50.0
django-request==1.5.6
djangorestframework==3.13.1
fake-useragent==0.1.11
gunicorn==20.1.0
idna==3.2
jsonfield==3.1.0
lxml==4.8.0
nose==1.3.7
oauthlib==3.2.0
psycopg2-binary==2.9.1
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
django-phone-field==1.8.1
Pillow==9.1.1